If supported avoid verbose mempool
we are interested only in mempool txids but the default call return much
more data.
The verbose parameter has been introduced in bitcoin from 0.25, on
elements we are still behind that.
We check once at startup if the feature is supported and use it if it
is.
Bitcoin mempool is usually much bigger than the elements one and I've
noticed in prod with 30k tx in mempool high usage of serde_json which
this will mitigate.