update time_since_last_block endpoint
This endpoint was returning a simple more or less than 10 minutes since
last block.
This is useful so that uptime checker can match a simple static  string
"less than 10 minutes" to provide insight if the service is running
correctly and up to date.

Ten minutes is not an unexpected time for bitcoin, so we adapted to 100
minutes to be strange for bitcoin (10 times to expected average).

Moreover we added the seconds elapsed so that client can potentially
parse and make their own logic.