fix bug, add skip_utxo
the commit to skip indexing outputs that not render as address
introduced a bug, because you can still spend script that are not
renderable as addresses, similar to the fact you can still spend an
empty script.
Took the opportunity to clearly define the two cases, if we want to skip
the utxo insertion, only when we are sure the script is unspendable and
when we want to skip indexing.